Split UMAPINFO code from g_mapinfo.* files into their own separate g_umapinfo.* files. Added some functionality to OScanner to prevent redundant code. (Functions that would be useful with OScanner elsewhere anyway)
