+++ common/g_umapinfo.h
@@ -14,13 +14,10 @@
 // GNU General Public License for more details.
 //
 // DESCRIPTION:
-//   Functions regarding reading and interpreting MAPINFO lumps.
+//   Functions regarding reading and interpreting UMAPINFO lumps.
 //
 //-----------------------------------------------------------------------------
 
 #pragma once
 
-extern BOOL HexenHack; // Semi-Hexen-compatibility mode
-
-void G_MapNameToLevelNum(level_pwad_info_t& info);
-void G_ParseMapInfo();
+void ParseUMapInfoLump(int lump, const char* lumpname);
